Refactor : Changed structure to have shared data across layers. #42
