+++ include/app/GuiLayer.hpp
@@ -1,12 +1,14 @@
 #pragma once
 #include <iostream>
-#include "external/raylib/raylib.h"
-#include "external/raylib/raygui.h"
-#include "core/Layer.hpp"
 #include <string>
 #include <vector>
 
-#include "core/Application.hpp"
+#include "external/raylib/raylib.h"
+#include "external/raylib/raygui.h"
+
+#include "app/Application.hpp"
+#include "app/Context.hpp"
+#include "core/Layer.hpp"
 
 /**
  * @struct Slider
@@ -26,27 +28,44 @@ struct slider
  * @class GUI layze
  * @brief The layer responsible for all the GUI informatations
  */
-class GUIlayer : public Core::Layer
+class GUILayer : public Core::Layer
 {
 public:
-    GUIlayer();
-    virtual ~GUIlayer();
+    GUILayer();
+    virtual ~GUILayer();
 
     virtual void update(double dt) override;
     virtual void render() override;
-    virtual void onEvent(Core::Event &event) {
-        std::cout << "Event recieved by the GUI"<< std::endl;    
-        }
-    virtual void stop();
+    virtual void onEvent(Core::Event &event) override;
+    virtual void stop() override;
 
+    /**
+     * @brief Called once initialized to link the app context (shared data) between all different layers
+     * @param context The context of the app
+     */
+    virtual void setContext(std::shared_ptr<Context> context) override { m_terrain = context->terrain; }
+
+    /**
+     * @brief apply all the changes of the sliders then dispach to the scene layer
+     */
     void applySliderChanges(std::string name, float value);
     void renderGui();
-    void toggleGUI();
-    bool getShowGUI();
+
+    /**
+     * @brief this alow to toggle on and off the GUI
+     */
+    void toggleGUI() { m_showGUI = !m_showGUI; }
+
+    /**
+     * @brief getter for the bool showGUI
+     * @return bool if we show the GUI or not
+     */
+    bool getShowGUI() const { return m_showGUI; }
     char seed[24] = "SEED";
     std::vector<slider> sliders;
 
 private:
+    std::shared_ptr<Core::Terrain> m_terrain; // Shared terrain from the app context
     bool m_textBoxEditMode;
 
     bool m_showGUI;
