+++ include/core/Event.hpp
@@ -1,5 +1,7 @@
 #pragma once
+
 #include <string>
+
 namespace Core
 {
   enum class EventType
