+++ src/app/SceneLayer.cpp
@@ -1,36 +1,31 @@
 #include <algorithm>
 #include <chrono>
 #include <iostream>
-#include "app/SceneLayer.hpp"
-#include "core/Application.hpp"
-#include "core/Terrain.hpp"
-#include "core/Event.hpp"
 
-SceneLayer::SceneLayer()
-{
-  m_camera = {{10.0f, 10.0f, 10.0f}, {0.0f, 0.0f, 0.0f}, m_cameraSpecs.upVector, m_cameraSpecs.fov, 0};
-  m_camera.target = m_terrain.getPos();
-}
+#include "app/Application.hpp"
+#include "app/SceneLayer.hpp"
+
+SceneLayer::SceneLayer() { m_camera = {{10.0f, 10.0f, 10.0f}, {0.0f, 0.0f, 0.0f}, m_cameraSpecs.upVector, m_cameraSpecs.fov, 0}; }
 SceneLayer::~SceneLayer() {}
 
 void SceneLayer::update(double dt)
 {
   m_handleInputs(dt);
   m_handleCamera(dt);
-  m_terrain.update(dt);
+  m_terrain->update(dt);
 }
 void SceneLayer::render()
 {
   BeginMode3D(m_camera);
 
   DrawGrid(50, 1.0);
-  m_terrain.render();
+  m_terrain->render();
 
   EndMode3D();
 }
 void SceneLayer::stop()
 {
-  m_terrain.unload();
+  m_terrain->unload();
 }
 
 void SceneLayer::m_handleInputs(double dt)
@@ -66,21 +61,21 @@ void SceneLayer::m_handleInputs(double dt)
       break;
     }
 
-    m_terrain.setSize(size);
-    m_terrain.generateCustomTerrain();
-    m_terrain.load();
-    std::cout << "INFO: Changed size to : " << m_terrain.getSideSize() << std::endl;
+    m_terrain->setSize(size);
+    m_terrain->generateCustomTerrain();
+    m_terrain->load();
+    std::cout << "INFO: Changed size to : " << m_terrain->getSideSize() << std::endl;
   }
 
   if (IsKeyPressed(KEY_Z))
-    m_terrain.switchRenderMode();
+    m_terrain->switchRenderMode();
 
   if (IsKeyPressed(KEY_T))
   {
     std::mt19937 gen(std::chrono::high_resolution_clock::now().time_since_epoch().count());
     std::uniform_int_distribution<unsigned int> distrib(0, 4294967295);
     unsigned int randomSeed = distrib(gen);
-    m_terrain.regenerateTerrain(randomSeed);
+    m_terrain->regenerateTerrain(randomSeed);
   }
 }
 
@@ -93,7 +88,7 @@ void SceneLayer::m_handleCameraInputs(double dt)
   if (IsKeyPressed(KEY_R))
     m_rotating = !m_rotating;
 
-  float r = Vector3Distance(m_camera.position, m_terrain.getPos());
+  float r = Vector3Distance(m_camera.position, m_terrain->getPos());
   float wheelZoom = GetMouseWheelMove() * dt * m_cameraSpecs.zoomSpeed;
 
   // If we are too close, we are compelled to zoom out or not zoom at all
@@ -122,47 +117,7 @@ void SceneLayer::m_handleCamera(double dt)
   if (m_camera.position.y > m_cameraSpecs.maxY)
     m_camera.position.y = m_cameraSpecs.maxY;
 
-  m_camera.target = m_terrain.getPos();
+  m_camera.target = m_terrain->getPos();
 }
 
-void SceneLayer::onEvent(Core::Event &event)
-{
-  std::cout << "Event on the scene layer" << std::endl;
-
-  std::cout << "EVENT Type : ";
-  switch (event.type)
-  {
-  case Core::EventType::CHANGE_NOISESCALE:
-    std::cout << "noisecale" << std::endl;
-    m_terrain.setNoiseAmplitude(event.value);
-    break;
-  case Core::EventType::CHANGE_PERSISTENCE:
-    std::cout << "Persitence" << std::endl;
-    m_terrain.setNoisePersistence(event.value);
-    break;
-  case Core::EventType::CHANGE_LACUNARITY:
-    std::cout << "Lacunarity" << std::endl;
-    m_terrain.setNoiseLacunarity(event.value);
-    break;
-  case Core::EventType::CHANGE_FREQUENCY:
-    std::cout << "Frequency" << std::endl;
-    m_terrain.setNoiseFrequency(event.value);
-    break;
-  case Core::EventType::CHANGE_OCTAVES:
-    std::cout << "Octaves" << std::endl;
-    m_terrain.setNoiseOctaves(event.value);
-    break;
-  case Core::EventType::CHANGE_SEED:
-    std::cout << "Seed" << std::endl;
-    m_terrain.regenerateTerrain((unsigned int)event.value);
-    break;
-  case Core::EventType::REGENERATE:
-    std::cout << "Regenerate" << std::endl;
-
-    break;
-  case Core::EventType::SETTINGS:
-    std::cout << "Settings" << std::endl;
-    break;
-  }
-  m_terrain.regenerateTerrain(123124);
-}
+void SceneLayer::onEvent(Core::Event &event) {}
